Move agent specific definitions to docker package
